% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSuitability.R
\name{calcSuitability}
\alias{calcSuitability}
\title{Calculating Suitability}
\usage{
calcSuitability(lu, spp, multipliers, landuseLayer = "melcc13")
}
\arguments{
\item{lu}{The study region land use raster (Raster)}

\item{spp}{List of species (list)}

\item{multipliers}{Stack of spatial multipliers (RasterStack)}

\item{landuseLayer}{The reference land use layer to link land use code to suitability (String)}
}
\value{
A stack of suitability rasters per species (0-100)
}
\description{
Estimate the suitability of study region for each species
}
