% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{classify_funcgroups}
\alias{classify_funcgroups}
\title{Title}
\usage{
classify_funcgroups(formatted_data, criterion)
}
\arguments{
\item{param1}{argument description}
}
\value{
what it returns
}
\description{
Description (paragraph)
}
\examples{
minimal reproducible eample

}
\seealso{
add link to additional resources if needed (or remove this line)
}
