% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{format_data}
\alias{format_data}
\title{Title}
\usage{
format_data(rawdata, focus_area)
}
\arguments{
\item{param1}{argument description}
}
\value{
what it returns
}
\description{
Description (paragraph)
}
\examples{
minimal reproducible eample

}
\seealso{
add link to additional resources if needed (or remove this line)
}
