% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse.r}
\alias{landuse.r}
\title{Land Use Dataset}
\format{
A raster with land use codes between 1-13; use reference land use codes table
 to link to descriptions
}
\source{
Ville de Nicolet project (2021)
}
\usage{
landuse.r
}
\description{
An example land use raster obtained from Ville de Nicolet project
}
\keyword{datasets}
