% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lu.codes}
\alias{lu.codes}
\alias{landuse.r}
\title{Land Use Codes}
\format{
A .csv linking descriptions to land use codes
}
\source{
Habitat internal
}
\usage{
landuse.r
}
\description{
Lookup table linking land use codes to descriptions
}
\keyword{datasets}
