% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{make_grid}
\alias{make_grid}
\title{Title}
\usage{
make_grid(...)
}
\arguments{
\item{param1}{argument description}
}
\value{
what it returns
}
\description{
Description (paragraph)
}
\examples{
minimal reproducible eample

}
\seealso{
add link to additional resources if needed (or remove this line)
}
