% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{max_dbh}
\alias{max_dbh}
\title{Title}
\usage{
max_dbh()
}
\arguments{
\item{param1}{argument description}
}
\value{
what it returns
}
\description{
Description (paragraph)
}
\examples{
minimal reproducible eample

}
\seealso{
add link to additional resources if needed (or remove this line)
}
