% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_diversity.R
\name{optimize_diversity}
\alias{optimize_diversity}
\title{optimize_diversity}
\usage{
optimize_diversity(
  region,
  inventory,
  make_grid = TRUE,
  grid_dimension,
  add_trees,
  n_add_trees
)
}
\arguments{
\item{param1}{argument description}
}
\value{
what it returns
}
\description{
Description (paragraph)
}
\examples{
minimal reproducible eample

}
\seealso{
add link to additional resources if needed (or remove this line)
}
