% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateau.R
\docType{data}
\name{plateau}
\alias{plateau}
\title{Plateau Mont-Royal delimitation}
\format{
A simple feature collection with 1 feature and 9 fields:
\describe{
  \item{MUNID}{todo}
  \item{CODEID}{todo}
  \item{CODEMAMROT}{todo}
  \item{NOM}{todo}
  \item{TYPE}{todo}
  \item{ABREV}{todo}
  \item{NUM}{todo}
  \item{AIRE}{todo}
  \item{PERIM}{todo} 
}
}
\source{
todo
}
\usage{
plateau
}
\description{
A polygon object describing the limits of the Plateau Mont-Royal arrondissement in Montreal (QC).
}
\keyword{datasets}
