% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateau_inventory.R
\docType{data}
\name{plateau_inventory}
\alias{plateau_inventory}
\title{Plateau Mont-Royal trees inventory}
\format{
A simple feature collection with 17530 features and 20 fields:
\describe{
  \item{INV_TYPE}{todo}
  \item{EMP_NO}{todo}
  \item{ARROND}{todo}
  \item{ARROND_NOM}{todo}
  \item{Rue}{todo}
  \item{COTE}{todo}
  \item{No_civique}{todo}
  \item{Emplcmnt}{todo}
  \item{Coord_X}{todo} 
  \item{Coord_Y}{todo}
  \item{SIGLE}{todo}
  \item{Ess_latin}{todo}
  \item{Essence_fr}{todo}
  \item{ESS_ANG}{todo}
  \item{DHP}{todo}
  \item{Date_relev}{todo}
  \item{Date_plnt}{todo}
  \item{LOCAL}{todo}
  \item{CODE_PARC}{todo}
  \item{NOM_PARC}{todo}
}
}
\source{
todo
}
\usage{
plateau_inventory
}
\description{
An inventory of all the trees in the Plateau Mont-Royal arrondissement of Montreal (QC).
}
\keyword{datasets}
