% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatialMultipliers}
\alias{spatialMultipliers}
\title{Spatial Multiplier Stack}
\format{
A raster stack 10 layers:
}
\source{
Ville de Nicolet project (2021)
}
\usage{
spatialMultipliers
}
\description{
A spatial multiplier raster stack obtained from Ville de Nicolet project
}
\details{
\itemize{
  \item{road_major_dist}{Distance to major roads (m)}
  \item{road_minor_dist}{weight of the diamond, in carats (m)}
  \item{anthro_dist}{Distance to anthropogenic lulc class (m)}
  \item{for_dist}{Distance to forests (m)}
  \item{for_drainage}{Drainage class (categorical)}
  \item{for_dense}{Forest Density? (m)}
  \item{water_dist}{Distance to aquatique class (m)}
  \item{for_edge_dist}{Distance to forest edge (m)}
  \item{for_age}{Forest Age (ans)}
  \item{wetland_dist}{Distance to humide (m)}
  #' }
}
\keyword{datasets}
