% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpeciesHabSuitLookup}
\alias{SpeciesHabSuitLookup}
\title{suitLookup}
\format{
list of 15 dataframes:
\itemize{
  \item{anthro_dist}{- Suitability based on distance to major roads}
  \item{dispersal}{- ***}
  \item{for_age}{- Suitability based on forest age}
  \item{for_dense}{ - Suitability based on Forest Density?}
  \item{for_depot}{***}
  \item{for_dist}{- Suitability based on distance to forest}
  \item{for_drainage}{- Suitability based drainage class (categorical)}
  \item{for_edge_dist}{- Distance to forest edge (m)}
  \item{habitat}{- Habitat characteristics for Species}
  \item{matrix}{- ***}
  \item{road_major_dist}{- Suitability based on distance to major roads}
  \item{road_minor_dist}{- Suitability based on distance to minor roads}
  \item{species}{- Species lookup table}
  \item{water_dist}{- Suitability based on distance to aquatique class (m)}
  \item{wetland_dist}{- Suitability based on distance to humide (m)}
  #' }
}
\source{
Habitat internal
}
\usage{
SpeciesHabSuitLookup
}
\description{
Lookup tables for converting spatial multipliers to species suitability
}
\keyword{datasets}
